requirements.txt
Flask
requests
python-dotenv
googletrans==3.1.0a0